fix(zc)!: Tan truncated its angle to a whole radian outside replays
The non-replay Tan path cast its radian argument to int16_t before
calling std::tan, so the fractional part of every angle was discarded
(and angles outside +/-32767 wrapped): Tan(0.5) returned tan(0) == 0,
and Tan(1.7) returned tan(1) ~= 1.56 instead of ~= -7.70. Sin and Cos
were unaffected.

Regressed recently in 582c719656.

+++ src/base/zc_math.cpp
@@ -151,7 +151,7 @@ double Tan(double x)
 		return r;
 	}
 	else
-		return std::tan((int16_t)x);
+		return std::tan(x);
 }
